truncate st_ino the same way git does it

+++ src/index.cc
@@ -91,7 +91,8 @@ bool IsModified(const git_index_entry* entry, const struct stat& st, bool trust_
     LOG(DEBUG) << "Dirty candidate (modified): " << Print(entry->path) << ": " #field " "
 
 #ifndef GITSTATUS_BOGUS_INO
-  COND(ino, entry->ino == st.st_ino) << entry->ino << " => " << st.st_ino;
+  COND(ino, entry->ino == static_cast<std::uint32_t>(st.st_ino))
+      << entry->ino << " => " << static_cast<std::uint32_t>(st.st_ino);
 #endif
 
   COND(stage, GIT_INDEX_ENTRY_STAGE(entry) == 0) << "=> " << GIT_INDEX_ENTRY_STAGE(entry);
